pyproject.toml
[tool.poetry]
name = "automation-exercise"
version = "0.1.0"
description = ""
authors = ["UserAnn7 <103679685+UserAnn7@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
playwright = "^1.44.0"
pytest = "^8.2.0"
pytest-playwright = "^0.4.4"
pytest-html = "^4.1.1"
allure-pytest = "^2.13.2"
pytest-soft-assertions = "^0.1.2"
behave = "^1.3.0"
parse = "^1.20.2"
allure-behave = "^2.15.0"

[tool.poetry.group.dev.dependencies]
behave-html-formatter = "^0.9.10"

